+++ src/collective/cart/shopping/browser/miscellaneous.py
@@ -1,0 +1,8 @@
+from Products.Five.browser import BrowserView
+from collective.cart.shopping.interfaces import IArticle
+
+
+class Miscellaneous(BrowserView):
+
+    def is_article(self):
+        return IArticle.providedBy(self.context)